+++ gluecodium/src/test/resources/smoke/basic_types/output/cpp/src/smoke/StringsWithCstring.cpp
@@ -1,0 +1,12 @@
+// -------------------------------------------------------------------------------------------------
+//
+//
+// -------------------------------------------------------------------------------------------------
+#include "smoke/StringsWithCstring.h"
+namespace smoke {
+StringsWithCstring::StringsWithCstring() {
+}
+StringsWithCstring::~StringsWithCstring() {
+}
+const char* StringsWithCstring::CSTRING_CONSTANT = "foobar";
+}